Make parsing and saving `no_std` compatible
Parsing and saving heavily rely on `std::io`, so it is far from trivial
to make those `no_std` compatible. However one possibility is simply
moving all the I/O out of the parsing and require that the data we parse
is already loaded from the file system or network or whatever the source
may be into the memory. This not only gets rid of all the I/O in the
parsers, but also makes them less generic, so there's less
monomorphization necessary and allows them to be faster as they can just
slice the data that is being passed into them, requiring much less
copying. So overall they got a lot faster and are independent of any
operating system now. We needed to port `quick-xml` to work on slices.
It's not great that we basically vendor a whole XML library, but it's
honestly not even that much code.