Introduce a `util` module
We accumulated a lot of individual parsing utilities and other utilities
all over the library by now. This refactors them all into a single
`util` module. Additionally there's a few other small cleanups included
such as the removal of `AsRef<str>` for setters by introducing a
`PopulateString` trait that can transfer ownership properly. Also
`indexmap` is removed in favor of a `Vec` based map that we already used
on `no_std` targets. It's faster anyway and now also makes use of the
`PopulateString` trait for populating the keys.