Borrow as much as possible in the parsers
Turns out that `serde_json` actually allows you to borrow from the input
string that you are parsing. I was under the impression that this was
not possible, as there might be escaped strings. However by using
`Cow<'a, str>` you can borrow when it's possible and if it's an escaped
string it unescapes it into the owned variant of the `Cow`. This also
extends it to our own s-expression parser and does a few minor
improvements here and there such as using `extend` to parse the segments
in order to preallocate enough memory for all the segments.