Implement unstable `WASI` support
The `WebAssembly System Interface` (WASI) specifies a set of APIs that
allow running the WebAssembly file as if it was running in a full
operating system. The `WASI` specification is not stable yet, so `WASI`
support is considered unstable for now and needs to be activated by
activating the `unstable` feature. For now `stdout`, `stdin` and
`stderr` are not hooked up and the file system is empty. We can
incrementally explore adding files to the file system as well as hooking
up `stdout` / `stderr` (it's unclear whether they should just log or
allow controlling the runtime by printing commands to it).