Let the `Resource Allocator` decode the images (#562)
We pass the encoded image data into the resource allocator now instead
of trying to decode the image in the scene manager. This way the
renderer has the ability to support all sorts of image formats that we
might not support. For example, this would also allow the renderer to
decode GIFs itself and thus properly animate them.