Use cargo's new `--crate-type` parameter
With Rust 1.64, cargo has a new `--crate-type` parameter that allows
overriding a crate's crate type. We need this for our C API where we
can't compile it to both a static and a dynamic library on every
platform. We had a workaround, where we had two wrapper crates that
reexported the libary's API and compiled to the their respective crate
type. These are now gone as `--crate-type` solves the entire problem. On
top of that, there was a linking problem with our wrapper crates, where
when LTO was not active, the API may not be exported at all. So debug
builds of our C API would previously not work at all.