Fix Integer Overflow in `FuzzyList` (#569)
Turns out the code I ported for the `FuzzyList` essentially counts
consecutive ranges of characters that match in both strings. The way it
does this is by multiplying the score by 2 for each matching character.
So this quickly overflows "any" integer type. We could saturate, but
since the score is fairly rough anyway and the original implementation
is in JavaScript where floating point numbers are used anyway, we simply
do the same here and switch to f64. Sorting floating point numbers is of
course iffy, but we don't even have any `NaN` values here anyway, so we
can just use a simplified ordering.