Improve the Correctness & Performance of Time Formatting
This improves the correctness and the performance of the time formatting
by making use of the fact that the underlying times are represented as a
pair of integers (seconds and the subsecond nanoseconds). This way
instead of formatting the fractional part as a floating point number we
already have the fractional part in form of the nanoseconds. On top of
that we can improve the performance a little bit by removing the data
dependency we introduced by calculating the hours from the total minutes
(and so on) when we can just calculate the hours directly from the total
seconds. This allows for a little bit of instruction level parallelism
where the days, hours, minutes and seconds can all be calculated in
parallel.