Don't use Floats to Parse Times
We want to be able to perfectly roundtrip a time, and since we already
switched to using only integers when formatting a time, we are able to
store them in a splits file in full precision, down to the nanosecond.
We want to be able to parse them again without any loss of precision.
In the PRs for formatting a time we gained a bunch of speed by switching
to the integers. I haven't done any benchmarks here, but improving the
parsing speed isn't the point of this PR anyway.