Fix Formatting Negative Times
It turns out that the new formatting code accidentally relied on the
fact that the sign for both the seconds and the subsecond nanoseconds
are always the same... and they almost always are. The problem is that
unlike floating point numbers, integers can't represent a separate
positive and negative 0. So if you have 0 seconds, the sign of the
nanoseconds actually matters. So you always have to take its sign into
account as well.