Add Support for Key Modifiers to the Hotkeys
This adds support for key modifiers to the `livesplit-hotkey` crate.
Instead of specifying a `KeyCode`, you now specify a `Hotkey` which
consists of a `KeyCode` and a set of `Modifiers`. All the
implementations support modifiers. However while `wasm-web` and `macOS`
natively provide the state of the modifiers to us, the other platforms
manually track their state.