Switch to `base64-simd`
This should improve the performance of parsing and formatting Base64. My
machine unfortunately doesn't have AVX2, so while it made a difference,
it wasn't as drastic as it could've been. Base64 however is the slowest
part when parsing splits at the moment, unless images need to be
resized, so it makes sense to improve the performance by using SIMD.