Don't repeat hotkeys on Windows (#584)
Unlike all the other platforms, on Windows we need to track the state of
each key ourselves. We simply use a byte array here where each bit
represents the state of the key. If we get another key down event when
the key is already marked as being held down, then we ignore it.