Implement `SpeedRunIGT` Parsing (#591)
`SpeedRunIGT` is a Minecraft mod that brings a full timer with deep auto
splitting capabilities to Minecraft. They recently added support for
https://therun.gg and it seems like it's quite commonly used, probably
the most commonly used timer after LiveSplit at this point. So it makes
sense for us to be able to parse these splits.

I also snuck in a few cleanups along the way (some now possible because
of Rust 1.65).