Implement Resolving Key Codes on Linux
Since we brought back X11, we can now mostly just use the established
X11 connection to resolve the key codes. This however does not work with
the `evdev` implementation. The solution there is to lazily open a
connection to X11 when trying to resolve a key code and simply fail
resolving the key code when a connection can't be established, then as
before, falling back onto the US keyboard layout. An unfortunate side
effect of this all however is that resolving a key code is now stateful,
meaning that you always need a `Hook` to do so.