Implement Resolving Key Codes on the Web
It was a bit unclear how to deal with the fact that this is an
asynchronous API on the web. But fortunately the promise resolves
immediately on the next "tick". So as long as we are fine with not being
able to resolve key codes on the first tick, we can just pretend it's a
synchronous operation by firing off the promise early and then simply
using it's result in the resolve function. Firing off the promise early
however is not that simple either, or at least was. Now that resolving
key codes is tied to the `Hook`, the constructor can easily be this
early point in time that fires off the promise and then starting with
the next tick the resolve function should work perfectly fine.