Optionally trust the C API to always receive valid UTF-8
This adds a new feature to the C API that allows it to skip UTF-8
validation on strings passed to it. Depending on where the C API is used
from you may always be sure that valid UTF-8 strings are being passed,
so no additional validation should be necessary. On top of that since we
always know that the strings passed from JavaScript in the web are going
to be valid UTF-8, the validation is always skipped there.