Introduce a new `max-opt` Cargo Profile (#598)
This introduces a new Cargo profile called `max-opt` that does as many
optimizations as possible. Previously we used the `release` profile for
that, but doing all these optimizations rarely makes sense locally and
instead only when deploying on the CI. Additionally I've noticed that
reducing the `codegen-units` to 1 allows for even more optimizations,
but slows compilation down even further. For both of these reasons the
`release` profile is now back to its original configuration in Cargo and
the `max-opt` profile is used for CI.