Allow custom handling of auto splitter log messages (#603)
Previously we would always log the auto splitter messages with the `log`
crate, but this may not necessarily be wanted. We already have a `Timer`
trait that abstracts all the interactions of the auto splitter with the
timer, so we might as well allow custom handling of the log messages
there.