Allow the auto splitters to add settings (#606)
The auto splitters can now specify settings that are meant to be shown
to and modified by the user. For now there are only boolean values. Also
to keep things simple, the auto splitters are reloaded when the settings
are changed. When registering the settings their values are returned, so
there doesn't need to be a separate "reload" API for now. This however
means that whenever there is such a reloading API, the auto splitters
probably would have to get updated to properly get notified. This is
acceptable breakage for such an early API though.