Don't store the file path in the `Run`
It basically always makes more sense to track this in the actual
application. This has been an artifact from being ported from the
original LiveSplit since the very beginning, but was never actually used
by any frontend. The desktop version now also doesn't end up using it,
so it's time to finally remove it.