Implement more Detailed Timer Settings
This implements the following settings / features for the Detailed
Timer:
1. The ability to set / override the colors of the timer, segment timer,
   segment name, comparison names and comparison times.
2. The ability to turn off the gradient of the timer and segment timer.
3. The ability to set the accuracy of the comparison times.

This also fixes a bug where the scene manager did not take into account
the sizes of the timer and segment timer at all.

I also included a fix for BMP images. We previously didn't read the
dimensions of BMP images at all, because we would re-encode them as PNG
anyway. However we may also want to shrink their dimensions, which was
always skipped because the dimension was never read from the BMP file.