Allow checking current attempt for new best times (#628)
When resetting an attempt we need to be able to check if the current
attempt contains new best times, which can be either new best segment
times or a new Personal Best. The user can then be asked whether to
update the splits with these new best times when resetting or not.

This also additional conditionally compiles out `Timer_print_debug` when
targetting WebAssembly without an operating system. This should reduce
the size of the WebAssembly binary on the web.