Switch to `fontdb` for Font Loading
We previously used `font-kit` which uses various native APIs for finding
and matching all the fonts. It's much harder to compile, especially with
`cross` and also brings in lots of functionality that we don't even
need. `fontdb` is a much simpler crate written in pure Rust that brings
everything that we need. It's previous matching algorithm wasn't as good
as it didn't take Name ID 16 into account, but that's now fixed.