Update various dependencies and fix a bug
This updates all the outdated dependencies and fixes a bug where
setting the linked layout through the run's method would mark the run as
modified. This therefore was even the case right after parsing a run.
Additionally tiny-skia can render on big endian systems now. I'm using
this as an opportunity to run the tests on all systems in CI now that
might be able to run more tests now.