Add `OS` and `Arch` querying to the Auto Splitting Runtime
This adds two new functions to the auto splitting runtime that allow an
auto splitter to query the operating system and architecture of the
platform that the runtime is running on. This allows an auto splitter to
dynamically adapt to the different platforms, such that for example
Linux and Windows can both be supported by the same auto splitter.