Use `cosmic-text` as the text engine
This switches our text engine from the lower level crates `fontdb`,
`rustybuzz` and `ttf-parser` to the higher level `cosmic-text` crate,
which uses those crates internally, but handles more higher level things
such as font fallback and BIDI text layout for us. For now we don't use
its rasterization capabilities and instead keep providing the glyphs as
paths, but we might do so in the future. We also don't use its layouting
and instead only its shaping. It's unclear if there is much benefit to
it as we have some more control if we do the layouting ourselves.