Fix Hotkey Modifiers on macOS
It turns out that the modifier flags are actual provided as part of the
event flags of the key down event. So we don't need to listen for the
flags to change anymore. We also accidentally interpreted the key code
of the flags changed event, which was often 0 (maybe always?), causing
us to interpret the flags changed event as a press of the `A` key.

Co-authored-by: Ted Wollman <25165500+TheTedder@users.noreply.github.com>