Add support for querying memory ranges (#664)
This exposes new function to the auto splitters for querying memory
ranges. At the moment this includes querying the amount of memory pages
a process has at the moment of querying. Then with any index that is
below the queried count, the address, the size and flags of the memory
region at that index can be queried. The supported flags for now are:
read, write, execute and whether a file path is known for the memory
range.

This API is quite racy, but is a lot cheaper than creating a snapshot of
all the regions. It's fine for the API to be a bit racy as the auto
splitters generally will retry the operation in the next tick anyway.
Also we don't refresh the modules in any of the functions that query an
individual memory range, so it should be fine for the most part. We can
add stronger guarantees later on where you can either lock the module
list or we guarantee that during a single tick you can't observe the
list racing.

---------

Co-authored-by: Christopher Serr <christopher.serr@gmail.com>