Switch from `winapi` to `windows-sys` (#703)
The `winapi` crate is the traditional way for interacting with the
Windows API. However, nowadays Microsoft is providing a more up-to-date
crate called `windows-sys`. This replaces the `winapi` dependency with
`windows-sys`. The only real disadvantage is that they release major
versions more often, resulting in more duplicates in the `Cargo.lock`.