Add `Config` type to initialize Auto Splitters
Rust does not have optional parameters, but we'd like to add more and
especially optional parameters when initializing an auto splitter. We
now provide a non-exhaustive `Config` type that has suitable default
values, that all can be overridden. This also adds a new experimental
setting for specifying an "interpreter script path", which allows us to
experiment with auto splitters that by themselves implement a script
engine and load a script from the file system to run. This script to
load from the file system is communicated to the auto splitter via the
`SCRIPT_PATH` environment variable.