auto_splitting: Add requests for settings
This patch adds 4 requests to the auto-splitting feature so that the library users are able to manage the auto splitter settings:
- GetSettings: Gets all of the user settings that are meant to be shown to the runner
- GetSettingsValue(key: String): Gets the SettingValue of a specific setting
- SetSettingsValue(key: String, value: SettingValue): Sets the SettingValue of a specific setting
- ReloadScript: Helper request that will reload the current script while keeping the same SettingsStore

This adds 3 new possible errors to tell the user what could have went wrong while executing the requests.

It also re export some of the useful types for settings related things.

Lastly, there's a new accessor in the runtime to get the SettingsStore as mutable and UserSetting / UserSettingKind are now clonable so that we can pass them along with the requests.