Update `wasmtime` to `12.0.1` (#720)
We were stuck on an old version of `wasmtime` because they removed the
ability to mark the file system as read only. We now wrap the file
system implementation and simply error out on anything that would write
to it. This is potentially a little buggier, but they plan on bringing
the feature back eventually anyway, so this implementation should be
fine in the meantime. I'm doing this upgrade, because their master
branch contains a bunch of improvements that we would want, such as
`serde` being decoupled from its `derive` crate and the debug support
now finally working. Unfortunately those are not released yet, but at
least we can now easily upgrade `wasmtime` again.