Allow nonblocking access to Auto Splitting Runtime
In order to implement the settings more cleanly and allow the auto
splitters to read and modify them, we need to reduce any potential locks
that could lock up the auto splitter, or worse, the entire application.
This now introduces a "git like" model where the user settings and
settings store can be atomically stored and loaded. This allows the
entire `asr-debugger` to function without being locked up by a malicious
auto splitter now.