Add API for accessing / modify settings maps
This adds a new API that allows auto splitters to access the current
settings map and to modify it. For now only booleans settings are
supported, but this can very easily be extended to other types. The API
is designed so no large scale locking is necessary, so malicious or
buggy auto splitters can't lock up the application. However, in order
for the API to not cause race conditions all the values are based on the
copy-on-write principle. The values are cheaply cloned, allowing the
auto splitter to work on a seemingly owned copy of the settings map and
the values inside. A CAS algorithm can be used to then resolve any races
where two parts of the code make modifications to the settings map in
parallel.

Co-authored-by: AlexKnauth <alexander@knauth.org>