Add support for setting values of type i64 and f64 (#737)
This adds support for setting values of type i64 and f64. It's unclear
if we want to support unsigned integers or different sizes. [OBS's Data
Settings
API](https://docs.obsproject.com/reference-settings#set-functions),
which logically is used for the same purpose as our settings maps and
values, also only suports these two types. So it's probably fine for now
to only support those.