Allow the Windows hotkey callback to recurse
This allows the hotkey callback on Windows to recurse. Each callback
hook is responsible for calling the next hook. Usually this is some
other hook, but it seems like it's also possible that this is our hook.
In that case the RefCell previously panicked due to being locked twice.
We instead now release the lock before calling the next hook. Having our
callback called twice is possibly still an indication of a bug
(alternatively we could ignore that second call), but for now this at
least ensures that the program doesn't crash.