Add Choice Settings to the Auto Splitting Runtime
This allows auto splitters to define settings widgets where the user can
choose between a set of options. These settings widgets could then be
visualized as for example dropdowns or radio buttons. Each option has a
string based key that is stored in the settings map as the value of the
setting when it is chosen.