Improve Auto Splitting runtime in `livesplit-core` (#749)
The auto splitting  runtime that lives directly in `livesplit-core` used
to have a channel that communicates with the background thread where the
auto splitter is getting executed. This channel was also used for
accessing and modifying the settings. However the design for the
settings has improved a lot to the point where multiple threads can
interact with them entirely lock free. This change removes the channel
entirely and instead provides the lock free way of interacting with the
settings.

Additionally there's a separate change where the auto splitter can now
be precompiled and reinstantiated separately. This allows you to restart
an auto splitter without also having to recompile it.