Allow hotkeys to be consumed
This adds a new type to `livesplit-hotkey` that allows you to specify
what your preference is for the hotkeys. You can now specify that you
want the hotkeys to be consumed, which means that the hotkeys will not
be forwarded to the application that is in focus. This is currently not
implemented for all the platforms, but the basic API for it is here now.