Fix Hotkey Modifiers on X11
Turns out that X11 is very strict about the hotkey modifiers. They need
to match exactly, otherwise the hotkey is not recognized. The problem is
that there's various "lock" modifiers, such as the caps lock, num lock,
and scroll lock, that are always considered for this, but are not really
keys that you explicitly press and instead are active or inactive until
you toggle them. And they need to match exactly as well. Whoever thought
that's a good design should stop doing so immediately.