Implement Layout Backgrounds (#774)
This implements background for the layouts. This is a lot more involved
than what it would seem. Part of it is a huge refactoring that changes
all layout states to be truly absolute. Previously they only included
image changes whenever an image changes. We did this because we didn't
want to serialize all the image data on every single frame. The new
approach is to have a cache for the images and only serialize the image
IDs. The image IDs are SHA-256 hashes, which allow deduplicating the
images as well.