Implement SVG Renderer
We internally represent the scene as vector graphics. Usually we expect
a renderer to take these and render them out, but it's also possible to
just serialize them out as an SVG image. I've had various experimental
implementations for a while. I thought it's time to upstream a proper
full implementation. This is going to be really useful for exporting a
screenshot of your current splits in the highest quality possible.