Force disable kerning for monospace text
`cosmic-text` unfortunately has no way to turn off kerning yet. However,
it's essential for our monospaced text to not have kerning. This
implements a small workaround that simply looks up (and caches) the x
advance for each glyph and uses it to advance the x coordinate instead
of the x advance that comes out of the shaping.

This also improves performance because the times on the right side of
the layout are all supposed to be monospaced. However, the fact that
they didn't use kerning meant that their widths were constantly
changing, changing the amount of space that is available on the left
side for the text. This caused the text on the left to recalculate its
ellipsis overflow state, which often forced the entire background layer
to be rerendered.