Implement Web Renderer
This adds a new renderer for the web that renders to a canvas, which can
then be easily attached anywhere in the DOM with any size. This should
allow the web LiveSplit One to replace its React / HTML based renderer,
bringing the rendering more in line with the native renderer, with
features like arbitrary resizing and horizontal splits. Additionally
this should improve the performance a lot and help remove a lot of the
code from the web LiveSplit One.