Switch `wasm-bindgen` bindings to `Symbol.dispose` (#804)
There is a [JavaScript
proposal](https://github.com/tc39/proposal-explicit-resource-management)
for using `Symbol.dispose` as a way to automatically clean up resources.
You do that by declaring variables like so:
```js
using resource = new Resource();
```
The variable `x` will be automatically disposed when it goes out of
scope. We used to have our own `dispose` method and `with` helper method
that allows scoped access to the resource, before it gets cleaned up at
the end. TypeScript already fully implements and polyfills support for
it, so we can already fully switch over to it.