Implement an Event System (#808)
This adds an event system. The events usually come from the hotkey
system, an auto splitter, the UI, or through a network connection. The
UI usually provides the implementation for the so called event sink,
forwarding all the events to the actual timer. It is able to intercept
the events and for example ask the user for confirmation before applying
them. Other handling is possible such as automatically saving the splits
or notifying a server about changes happening in the run.