Increase PNG compression level (#817)
This commit increases the compression level for when we encode PNG
images. We do this for Llanfair files (Gered and original) were the
images are uncompressed and also for when we encounter BMP images in
LiveSplit files or we scale down images. Increasing the compression
level doesn't seem to really take much more time, but has a decent
impact on the size of the images.