Prevent `TimeStamp`s from being modified (#821)
Our `TimeStamp` type is internally an `Instant`. The representation of
`Instant` depends on the platform, but for many platforms it internally
consists of unsigned numbers. For WASI its worse, because the moment the
WASI VM starts, the time starts at 0. This means any subtraction from
the time results in a numeric underflow and might panic. To prevent this
from happening, `TimeStamp`s can no longer be manually modified, they
always represent valid points in time.