Added runtime_get current_tick
This adds a new tick field that increases by 1 at every call of the update function. This can be used for different purposes, as in for caching data or to check (together with the current tick rate, albeit in an approximate way), the amount of time passed when executing code between iterations of the main autosplitting loop.