Update to wasmtime 22 (#824)
This updates us to the latest version of wasmtime. The WASI API changed
a lot, because WASI 0.2 got released since then. We still only support
WASI 0.1 until there's practical toolchains for targeting WASI 0.2, so
we can experiment with it.

Unfortunately the new API does not allow us to define custom file system
logic anymore, so mapping to Windows device paths is not possible
anymore.