Add `stderr` Support for Auto Splitters
Auto splitters can now print to `stderr`, which is the same as printing
via `print_message`, except that like a usual terminal it is line
buffered and either flushed on new lines or manually. This then ends up
in the same `log_auto_splitter` call that `print_message` ends up in.
The logging now also differentiates between messages that come from the
auto splitter and messages that come from the runtime. Additionally this
bumps `wasmtime` to version 24.