Fix panic when formatting `TimeStamp::MIN`
In order to print the minus sign of times, we detect the need for the
minus sign, print it and then print the absolute value of the time.
However, integers have a minimum value that is one less than the
absolute value of the maximum value. This means that the minimum value
of a signed integer cannot be negated. This caused a panic when
formatting `TimeStamp::MIN` when overflow checks are enabled, such as
when compiling with the debug profile.

There was also a big lack of tests for the formatting of times. Tests
are now included for all the important cases.