Fix panic when formatting `TimeStamp::MIN` (#838)
In order to print the minus sign of times, we detect the need for the
minus sign, print it and then print the absolute value of the time.
However, integers have a minimum value of which the absolute value is
larger than the maximum value of the integer. So for example `i8` has a
minimum value of `-128` and a maximum value of `127`. The absolute value
therefore can't be represented. This caused a panic when formatting
`TimeStamp::MIN` when overflow checks are enabled, such as when
compiling with the debug profile.

There was also a big lack of tests for the formatting of times. Tests
are now included for all the important cases.